requirements.txt
flask==3.0.3
pyotp==2.9.0
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.32.3
neo_api_client @ git+https://github.com/Kotak-Neo/Kotak-neo-api-v2.git@v2.0.1
